Fix: Error in GPIO driver
Changes the data type input of the GPIO_WritePort() function,
uint8_t to _uint16_t.

This error cause a bug at write in a port, writing only the
the first 8 pines of the port.

+++ Inc/stm32f103_gpio_driver.h
@@ -124,7 +124,7 @@ uint16_t GPIO_ReadPort(GPIO_RegDef_t *GPIOx);
  * Pheriperal WRITE
  */
 void GPIO_WitePin(GPIO_RegDef_t *GPIOx, uint8_t PinNumber, uint8_t Value);
-void GPIO_WritePort(GPIO_RegDef_t *GPIOx, uint8_t Value);
+void GPIO_WritePort(GPIO_RegDef_t *GPIOx, uint16_t Value);
 void GPIO_TogglePin(GPIO_RegDef_t *GPIOx, uint8_t PinNumber);
 
 /*
